requirements.txt
aws-cdk-lib>=2.100.0
constructs>=10.0.0
boto3>=1.28.0
strands-agents>=0.1.0
pydantic>=2.0.0
python-docx>=0.8.11
openpyxl>=3.1.0
replicate>=0.25.0
requests>=2.31.0
pytest>=7.4.0
pytest-mock>=3.12.0
moto>=4.2.0
